1 初步完成客户端代码重构 2 修复了一个启动bug 3 修复两个内存泄漏问题
ps：发现多线程发送命令冲突的问题

+++ RemoteClient/ClientSocket.h
@@ -169,11 +169,11 @@ public:
 #define BUFFER_SIZE 2048000
 	int DealCommand() {
 		if (m_sock == -1)return -1;
-		char* buffer = m_buffer.data();
+		char* buffer = m_buffer.data(); // TODO：多线程发送命令时可能会冲突
 		static size_t index = 0;
 		while (true) {
 			size_t len = recv(m_sock, buffer + index, BUFFER_SIZE - index, 0);
-			if ((len <= 0) && (index <= 0)) {
+			if (((int)len <= 0) && ((int)index <= 0)) {
 				return -1;
 			}
 			//Dump((BYTE*)buffer, index);
